Instead of @__rdl_inst for type-instantiated objects, use a more general @__rdl_type field.
Add check_member_or_leq(obj) method that checks that obj either has a @__rdl_type field that's <= self, or is a member or self, and throws an exception if not.

Use check_member_or_leq in type contracts and in checking method calls.

Also report correct position of ill-typed arguments.